refactor(review): move field-change projection from application to domain (E1)
project_changed, project_relation, project_impacts are pure functions
over a ChangedObject — no I/O, no orchestration. They were stranded in
application/review.rs because the V3-DESIGN public-API listing put
field_changes() there, but the function is a domain projection over the
FieldChange vocabulary it produces.

Move project_changed (and its private helpers) into a new
domain/review/projection.rs. Move the ~25 inline projection tests
(field_changes_projection module, ~435 lines) with the code per ADR-0008
— tests travel with the white-box contract they exercise.

application/review.rs shrinks from 1780 → 1142 lines. diff_objects
becomes a four-line orchestration that calls into the domain
projection. Wire envelope and aggregator code stays in application
(addressed in subsequent E-steps).
